requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
cors==1.0.1
Django==5.1.1
django-cors-headers==4.4.0
django-unfold==0.39.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filelock==3.16.1
future==1.0.0
gevent==24.2.1
greenlet==3.1.1
idna==3.10
inflection==0.5.1
numpy==2.1.1
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pycparser==2.22
PyJWT==2.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
requests==2.32.3
requests-file==2.1.0
six==1.16.0
sqlparse==0.5.1
tldextract==5.1.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
zope.event==5.0
zope.interface==7.0.3
